requirements - Copy.txt
absl-py==1.3.0
aiosignal==1.3.1
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
black==22.12.0
blinker==1.5
bs4==0.0.1
bullet==2.2.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.4
cloudpickle==1.6.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
daemoniker==0.2.3
dask==2023.1.0
decorator==5.1.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@650c60ed365db1b04765ffd121574c0765a5c24c
dill==0.3.6
distlib==0.3.6
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
fastapi==0.89.0
ffmpeg-python==0.2.0
filelock==3.9.0
fire==0.5.0
flatbuffers==23.1.4
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h11==0.14.0
h5py==3.7.0
huggingface-hub==0.12.0
humanfriendly==10.0
hydra-core==1.3.1
idna==3.4
img2table==0.0.10
importlib-metadata==6.0.0
imutils==0.5.4
inflection==0.5.1
iniconfig==2.0.0
iopath==0.1.9
ipython==8.8.0
jax==0.4.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.11.0
keras-nightly==2.12.0.dev2023010808
kiwisolver==1.4.4
layoutparser==0.3.4
libclang==14.0.6
llvmlite==0.39.1
locket==1.0.0
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
msgpack==1.0.4
mypy-extensions==0.4.3
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.3
parso==0.8.3
partd==1.3.0
pathspec==0.10.3
pdf2image==1.16.2
pdfminer.six==20221105
pdfplumber==0.7.6
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
pluggy==1.0.0
polars==0.16.2
portalocker==2.7.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.6
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.4
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
PyMuPDF==1.21.1
pyparsing==3.0.9
pyreadline3==3.4.1
Pyro4==4.82
pyrsistent==0.19.3
pytesseract==0.3.10
pytest==7.2.1
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==305
PyYAML==6.0
ray==2.2.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.3.1
rsa==4.9
sacremoses==0.0.53
sahi==0.11.11
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
semver==2.13.0
sentencepiece==0.1.97
sentry-sdk==1.15.0
serpent==1.41
shap==0.41.0
shapely==2.0.1
simple-term-menu==1.6.1
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
starlette==0.22.0
streamlit==1.18.0
tabulate==0.9.0
tb-nightly==2.12.0a20230107
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
terminaltables==3.1.10
tf-estimator-nightly==2.12.0.dev2023010809
tf-nightly==2.12.0.dev20230108
tf_nightly_intel==2.12.0.dev20230108
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
transformers==4.25.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ultralytics==8.0.21
ultralyticsplus==0.0.23
urllib3==1.26.13
uvicorn==0.20.0
validators==0.20.0
virtualenv==20.19.0
Wand==0.6.11
watchdog==2.2.1
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.14.1
XlsxWriter==3.0.8
xmltodict==0.12.0
yacs==0.1.8
zerocopy==0.1.0
zipp==3.11.0
